requirements.txt
# =============================================================================
# Credit Scoring Knowledge Distillation Framework
# Requirements
# =============================================================================

# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Machine Learning
scikit-learn>=1.0.0
xgboost>=1.7.0
lightgbm>=3.3.0
catboost>=1.0.0

# Data Processing
pandas>=1.5.0
numpy>=1.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Interpretability
shap>=0.41.0

# Hyperparameter Optimization
optuna>=3.0.0

# Experiment Tracking (optional)
# mlflow>=2.0.0

# Configuration
pyyaml>=6.0

# Utilities
tqdm>=4.64.0
openpyxl>=3.0.0

# Jupyter Support (optional)
# jupyter>=1.0.0
# ipykernel>=6.0.0
